pyproject.toml
[tool.poetry]
name = "carlet-scraper"
version = "0.1.0"
description = ""
authors = ["asdf50003206 <asdf500032061@gmail.com>"]
readme = "README.md"
packages = [{include = "carlet_scraper"}]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.12.0"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.7"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pillow = "^10.1.0"
pytesseract = "^0.3.10"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
